API: Reformulate cufft.Ptycho as composition of Propagate and Convolution
Because:
- The ptychography operator is naturally composed of a propagation
and convolution operations
- These two new operators are required for the ADMM formulation of
ptychography
- The convolution operator can be reused for solving other problems

This refactoring implements two new C++ operators for just doing
far-field propagation with an FFT and linear convolution in the
near-field. These operators will accept numpy arrays and move
data to the GPU internally. This means that the client library
does not have to use CuPy arrays or worry about memory limits on
the GPU.