API: Move padding to diffraction operator and limit to 1 mode
The padding operation was done in the propagation operator, but
here it is moved to the diffraction operator because it is more
memory efficient to allocate the nearplane patches when
extracting them instead of padding later before the FFT. This
also makes the propagation operator simpler because it doesn't
contain padding or clipping operators and this operation may
be done in-place.

In order to save memory, the operators are now also limited to
1 incorherent mode at a time. This is because storing the wave
fronts for 20+ modes simultaneously takes a lot of memory and
for GPUs we will run out. Looping through the modes in the code
adds cognitive and computational overhead (python loops are
slow), but the runtime memory saving are worth it.

An overwrite boolean parameter has also been added to some
function signatures to signify that the input arrays may be over
written for parameters whose input arrays match output arrays
size.